+++ src/main/java/com/sparta/hotitemcollector/domain/user/dto/auth/RefreshRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.hotitemcollector.domain.user.dto;
+package com.sparta.hotitemcollector.domain.user.dto.auth;
 
 import lombok.Getter;
